+++ relay-server/src/utils/rate_limits.rs
@@ -783,6 +783,7 @@ impl<F> fmt::Debug for EnvelopeLimiter<F> {
 mod tests {
     use std::collections::{BTreeMap, BTreeSet};
 
+    use relay_base_schema::organization::OrganizationId;
     use relay_base_schema::project::{ProjectId, ProjectKey};
     use relay_metrics::MetricNamespace;
     use relay_quotas::RetryAfter;
@@ -803,7 +804,7 @@ mod tests {
         // Add a generic rate limit for all categories.
         rate_limits.add(RateLimit {
             categories: DataCategories::new(),
-            scope: RateLimitScope::Organization(42),
+            scope: RateLimitScope::Organization(OrganizationId::new(42)),
             reason_code: Some(ReasonCode::new("my_limit")),
             retry_after: RetryAfter::from_secs(42),
             namespaces: smallvec![],
@@ -830,7 +831,7 @@ mod tests {
         // Rate limit with reason code and namespace.
         rate_limits.add(RateLimit {
             categories: smallvec![DataCategory::MetricBucket],
-            scope: RateLimitScope::Organization(42),
+            scope: RateLimitScope::Organization(OrganizationId::new(42)),
             reason_code: Some(ReasonCode::new("my_limit")),
             retry_after: RetryAfter::from_secs(42),
             namespaces: smallvec![MetricNamespace::Custom, MetricNamespace::Spans],
@@ -839,7 +840,7 @@ mod tests {
         // Rate limit without reason code.
         rate_limits.add(RateLimit {
             categories: smallvec![DataCategory::MetricBucket],
-            scope: RateLimitScope::Organization(42),
+            scope: RateLimitScope::Organization(OrganizationId::new(42)),
             reason_code: None,
             retry_after: RetryAfter::from_secs(42),
             namespaces: smallvec![MetricNamespace::Spans],
@@ -854,7 +855,7 @@ mod tests {
     #[test]
     fn test_parse_invalid_rate_limits() {
         let scoping = Scoping {
-            organization_id: 42,
+            organization_id: OrganizationId::new(42),
             project_id: ProjectId::new(21),
             project_key: ProjectKey::parse("a94ae32be2584e0bbd7a4cbb95971fee").unwrap(),
             key_id: Some(17),
@@ -868,7 +869,7 @@ mod tests {
     #[test]
     fn test_parse_rate_limits() {
         let scoping = Scoping {
-            organization_id: 42,
+            organization_id: OrganizationId::new(42),
             project_id: ProjectId::new(21),
             project_key: ProjectKey::parse("a94ae32be2584e0bbd7a4cbb95971fee").unwrap(),
             key_id: Some(17),
@@ -885,7 +886,7 @@ mod tests {
             vec![
                 RateLimit {
                     categories: DataCategories::new(),
-                    scope: RateLimitScope::Organization(42),
+                    scope: RateLimitScope::Organization(OrganizationId::new(42)),
                     reason_code: Some(ReasonCode::new("my_limit")),
                     retry_after: rate_limits[0].retry_after,
                     namespaces: smallvec![],
@@ -911,7 +912,7 @@ mod tests {
     #[test]
     fn test_parse_rate_limits_namespace() {
         let scoping = Scoping {
-            organization_id: 42,
+            organization_id: OrganizationId::new(42),
             project_id: ProjectId::new(21),
             project_key: ProjectKey::parse("a94ae32be2584e0bbd7a4cbb95971fee").unwrap(),
             key_id: Some(17),
@@ -925,7 +926,7 @@ mod tests {
             rate_limits,
             vec![RateLimit {
                 categories: smallvec![DataCategory::MetricBucket],
-                scope: RateLimitScope::Organization(42),
+                scope: RateLimitScope::Organization(OrganizationId::new(42)),
                 reason_code: None,
                 retry_after: rate_limits[0].retry_after,
                 namespaces: smallvec![MetricNamespace::Custom, MetricNamespace::Spans],
@@ -936,7 +937,7 @@ mod tests {
     #[test]
     fn test_parse_rate_limits_empty_namespace() {
         let scoping = Scoping {
-            organization_id: 42,
+            organization_id: OrganizationId::new(42),
             project_id: ProjectId::new(21),
             project_key: ProjectKey::parse("a94ae32be2584e0bbd7a4cbb95971fee").unwrap(),
             key_id: Some(17),
@@ -951,7 +952,7 @@ mod tests {
             rate_limits,
             vec![RateLimit {
                 categories: smallvec![DataCategory::MetricBucket],
-                scope: RateLimitScope::Organization(42),
+                scope: RateLimitScope::Organization(OrganizationId::new(42)),
                 reason_code: Some(ReasonCode::new("some_reason")),
                 retry_after: rate_limits[0].retry_after,
                 namespaces: smallvec![],
@@ -962,7 +963,7 @@ mod tests {
     #[test]
     fn test_parse_rate_limits_only_unknown() {
         let scoping = Scoping {
-            organization_id: 42,
+            organization_id: OrganizationId::new(42),
             project_id: ProjectId::new(21),
             project_key: ProjectKey::parse("a94ae32be2584e0bbd7a4cbb95971fee").unwrap(),
             key_id: Some(17),
@@ -976,7 +977,7 @@ mod tests {
             rate_limits,
             vec![RateLimit {
                 categories: smallvec![DataCategory::Unknown, DataCategory::Unknown],
-                scope: RateLimitScope::Organization(42),
+                scope: RateLimitScope::Organization(OrganizationId::new(42)),
                 reason_code: None,
                 retry_after: rate_limits[0].retry_after,
                 namespaces: smallvec![],
@@ -1018,7 +1019,7 @@ mod tests {
     fn rate_limit(category: DataCategory) -> RateLimit {
         RateLimit {
             categories: vec![category].into(),
-            scope: RateLimitScope::Organization(42),
+            scope: RateLimitScope::Organization(OrganizationId::new(42)),
             reason_code: None,
             retry_after: RetryAfter::from_secs(60),
             namespaces: smallvec![],